Move common metrics into internal metrics
